PIM-3949: re-work the abstract processor
