+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/AbstractProcessor.php
@@ -18,7 +18,7 @@ use Symfony\Component\Validator\ValidatorInterface;
 /**
  * Abstract processor to provide a way to denormalize array data to object by,
  * - fetch an existing object or create it
- * - denormalize item to update the object
+ * - update the object
  * - validate the object
  * - skip the object if it contains invalid data
  *
@@ -33,40 +33,22 @@ abstract class AbstractProcessor extends AbstractConfigurableStepElement impleme
     /** @var StepExecution */
     protected $stepExecution;
 
-    /** @var ValidatorInterface */
-    protected $validator;
-
     /** @var IdentifiableObjectRepositoryInterface */
     protected $repository;
 
-    /** @var DenormalizerInterface */
-    protected $denormalizer;
-
-    /** @var ObjectDetacherInterface */
-    protected $detacher;
-
-    /** @var string */
-    protected $class;
+    /** @var ValidatorInterface */
+    protected $validator;
 
     /**
      * @param IdentifiableObjectRepositoryInterface $repository   repository to search the object in
-     * @param DenormalizerInterface                 $denormalizer denormalizer used to transform array to object
      * @param ValidatorInterface                    $validator    validator of the object
-     * @param ObjectDetacherInterface               $detacher     object detacher
-     * @param string                                $class        class of the object to instanciate in case if need
      */
     public function __construct(
         IdentifiableObjectRepositoryInterface $repository,
-        DenormalizerInterface $denormalizer,
-        ValidatorInterface $validator,
-        ObjectDetacherInterface $detacher,
-        $class
+        ValidatorInterface $validator
     ) {
         $this->repository = $repository;
-        $this->denormalizer = $denormalizer;
         $this->validator = $validator;
-        $this->detacher = $detacher;
-        $this->class = $class;
     }
 
     /**
@@ -83,26 +65,6 @@ abstract class AbstractProcessor extends AbstractConfigurableStepElement impleme
     public function setStepExecution(StepExecution $stepExecution)
     {
         $this->stepExecution = $stepExecution;
-    }
-
-    /**
-     * Try to find an object according to its identifiers from a repository or create an empty object
-     * if it does not exist.
-     *
-     * @param IdentifiableObjectRepositoryInterface $repository the repository to search inside
-     * @param array                                 $data       the data that is currently processed
-     * @param string                                $class      the class to instanciate in case the
-     *                                                          object has not been found
-     *
-     * @return object
-     */
-    protected function findOrCreateObject(IdentifiableObjectRepositoryInterface $repository, array $data, $class)
-    {
-        if (null !== $object = $this->findObject($repository, $data)) {
-            return $object;
-        }
-
-        return new $class();
     }
 
     /**
@@ -128,19 +90,6 @@ abstract class AbstractProcessor extends AbstractConfigurableStepElement impleme
         }
 
         return $repository->findOneByIdentifier(implode('.', $references));
-    }
-
-    /**
-     * Detaches the object from the unit of work
-     *
-     * Detach an object from the UOW is the responsibility of the writer, but to do so, it should know the
-     * skipped items or we should use an explicit persist strategy
-     *
-     * @param mixed $object
-     */
-    protected function detachObject($object)
-    {
-        $this->detacher->detach($object);
     }
 
     /**
