+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/AttributeOptionProcessor.php
@@ -17,7 +17,7 @@ use Symfony\Component\Validator\ValidatorInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AttributeOptionProcessor extends AbstractReworkedProcessor
+class AttributeOptionProcessor extends AbstractProcessor
 {
     /** @var StandardArrayConverterInterface */
     protected $arrayConverter;
