+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/VariantGroupProcessor.php
@@ -38,6 +38,12 @@ class VariantGroupProcessor extends AbstractProcessor
     /** @staticvar string */
     const LABEL_PATTERN = 'label-';
 
+    /** @var DenormalizerInterface */
+    protected $denormalizer;
+
+    /** @var ObjectDetacherInterface */
+    protected $detacher;
+
     /** @var NormalizerInterface */
     protected $normalizer;
 
@@ -49,6 +55,9 @@ class VariantGroupProcessor extends AbstractProcessor
 
     /** @var string */
     protected $format;
+
+    /** @var string */
+    protected $class;
 
     /**
      * @param IdentifiableObjectRepositoryInterface $groupRepository
@@ -72,11 +81,14 @@ class VariantGroupProcessor extends AbstractProcessor
         $templateClass,
         $format
     ) {
-        parent::__construct($groupRepository, $denormalizer, $validator, $detacher, $groupClass);
+        parent::__construct($groupRepository, $validator);
+        $this->denormalizer         = $denormalizer;
+        $this->detacher             = $detacher;
         $this->normalizer           = $normalizer;
         $this->templateMediaManager = $templateMediaManager;
         $this->templateClass        = $templateClass;
         $this->format               = $format;
+        $this->class                = $groupClass;
     }
 
     /**
@@ -102,7 +114,10 @@ class VariantGroupProcessor extends AbstractProcessor
      */
     protected function findOrCreateVariantGroup(array $groupData)
     {
-        $variantGroup = $this->findOrCreateObject($this->repository, $groupData, $this->class);
+        if (null === $variantGroup = $this->findObject($this->repository, $groupData)) {
+            $variantGroup = new $this->class();
+        }
+
         $isExistingGroup = (null !== $variantGroup->getType() && false === $variantGroup->getType()->isVariant());
         if ($isExistingGroup) {
             $this->skipItemWithMessage(
@@ -257,5 +272,18 @@ class VariantGroupProcessor extends AbstractProcessor
         }
 
         return $template;
+    }
+
+    /**
+     * Detaches the object from the unit of work
+     *
+     * Detach an object from the UOW is the responsibility of the writer, but to do so, it should know the
+     * skipped items or we should use an explicit persist strategy
+     *
+     * @param mixed $object
+     */
+    protected function detachObject($object)
+    {
+        $this->detacher->detach($object);
     }
 }
