+++ src/Pim/Bundle/BaseConnectorBundle/Writer/Doctrine/BaseWriter.php
@@ -37,9 +37,8 @@ class BaseWriter extends AbstractConfigurableStepElement implements
         BulkSaverInterface $bulkSaver,
         ObjectDetacherInterface $objectDetacher
     ) {
-        $this->bulkSaver = $bulkSaver;
+        $this->bulkSaver      = $bulkSaver;
         $this->objectDetacher = $objectDetacher;
-        $this->total = 0; // TODO: to drop, debug purpose
     }
 
     /**
@@ -47,15 +46,8 @@ class BaseWriter extends AbstractConfigurableStepElement implements
      */
     public function write(array $objects)
     {
-        $this->total = count($objects) + $this->total;
-        echo $this->convert(memory_get_usage(true)).' write '.count($objects).'/'.$this->total.PHP_EOL;
-
         $this->bulkSaver->saveAll($objects);
-
-        // TODO: we should always detach objects we just processed, unfortunately, we encounter issues with flush()
-        // and fact that some subscribers as AddVersionListener loads others objects, for instance, when import a new
-        // attribute option, we detach the option, which is reloaded to handle the versioning of an attribute and
-        // inserted twice, with DEFERRED_EXPLICIT it works well !!!
+        // TODO a detachAll could be nice!
         foreach ($objects as $object) {
             $this->objectDetacher->detach($object);
         }
@@ -75,18 +67,5 @@ class BaseWriter extends AbstractConfigurableStepElement implements
     public function setStepExecution(StepExecution $stepExecution)
     {
         $this->stepExecution = $stepExecution;
-    }
-
-    /**
-     * TODO: to drop, debug purpose
-     *
-     * @param $size
-     * @return string
-     */
-    public function convert($size)
-    {
-        $unit=array('b','kb','mb','gb','tb','pb');
-
-        return @round($size/pow(1024,($i=floor(log($size,1024)))),2).' '.$unit[$i];
     }
 }
